+++ src/models/schedule.js
@@ -21,7 +21,7 @@ module.exports = (sequelize, DataTypes) => {
   Schedule.init({
     currentNumber: DataTypes.INTEGER,
     maxNumber: DataTypes.INTEGER,
-    date: DataTypes.INTEGER,
+    date: DataTypes.BIGINT,
     timeType: DataTypes.STRING,
     doctorId: DataTypes.INTEGER,
     hasBooking: DataTypes.BOOLEAN,
